convert to float vector

+++ Final/Body.cpp
@@ -1,6 +1,6 @@
 #include "Body.h"
 
-Body::Body(Vector3d size) : Rect(size)
+Body::Body(Vector3f size) : Rect(size)
 {
 }
